Runs, doesn't return correct values
TwoDimMSM works fine using pseudoAttraction as a forcing function. This pushes the masses toward the center of mass in the 2d system.

ThreeDimMSM has an issue with a migrating center of mass that needs to be sorted out and is very much a work in progress. I can use it to make compressed 3d systems by keeping the max time very short. What I really want is to reach a stable equilibrium so I know everything is working. Instead we see a system migration apparently with all of the masses moving in the same direction.